requirements.txt
amqp==2.3.2
anyjson==0.3.3
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cython==0.29.2
dj-database-url==0.5.0
django-admin-sortable2==0.7
django-appconf==1.0.2
django-cache-url==3.0.0
django-celery-email==2.0.1
django-classy-tags==0.8.0
django-cms==3.6.0
django-configurations==2.1
django-contrib-comments==1.9.0
django-filer==1.4.1
django-filter==2.0.0
django-floppyforms==1.7.0
django-formtools==2.1
django-fsm==2.6.0
django-elasticsearch-dsl==0.5.1
django-leaflet==0.24.0
django-markdown-deux==1.0.5
django-mptt==0.9.1
django-newsletter==0.8b1
django-oauth-toolkit==1.2.0
django-parler==1.9.2
django-polymorphic==2.0.3
django-sekizai==0.10.0
django-storages==1.7.1
django-taggit==0.23.0
django-taggit-autosuggest==0.3.4
django-templatetag-sugar==1.0
django-tinymce==2.7.0
django-treebeard==4.3
django-wikidata==0.0.6
django<2.2
djangocms-admin-style==1.2.9
djangocms-apphook-setup==0.3.0
djangocms-attributes-field==1.0.0
djangocms-picture==2.1.3
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
djangorestframework-csv==2.1.0
djangorestframework-jsonp==1.0.2
djangorestframework==3.9.0
easy-thumbnails==2.5
elasticsearch==6.3.1
# Keep back for django-elasticsearch-dsl
elasticsearch-dsl==6.1.0
geocoder==1.38.1
geoip2==2.9.0
geopy==1.18.1
honcho==1.0.1
html5lib==0.999999999
icalendar==4.0.3
idna==2.7
itypes==1.1.0
jinja2==2.10
jsonfield==2.0.2
kombu==4.3.0
lxml==4.2.5
markdown2==2.3.7
markdown==3.0.1
markupsafe==1.1.0
maxminddb==1.4.1
oauthlib==2.1.0
pdflib==0.1.7
phonenumbers==8.10.2
pillow==5.3.0
psycopg2-binary==2.7.6.1
pyisemail==1.3.2
pylatex==1.3.0
pypdf2==1.26.0
python-magic==0.4.15
python-memcached==1.59
python-mimeparse==1.6.0
python-slugify==1.2.6
pytz==2018.7
pyyaml==3.13
raven==6.9.0
reportlab==3.5.12
requests==2.21.0
unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
urlobject==2.4.3
vine==1.1.4
wand==0.4.5
yurl==0.13
django-extensions==2.1.5

# -e git+git://github.com/sirfz/tesserocr.git@35ae0f535f3d33679b3b93d7fc4e08ab2e6a92a1#egg=tesserocr

-e git+git://github.com/okfde/froide.git@master#egg=froide
-e git+git://github.com/okfde/froide-campaign.git@master#egg=froide-campaign
-e git+git://github.com/okfde/froide-food.git@master#egg=froide-food
-e git+git://github.com/okfde/froide-fax.git@master#egg=froide-fax
-e git+git://github.com/okfde/froide-legalaction.git@master#egg=froide-legalaction
-e git+git://github.com/okfde/froide-exam.git@master#egg=froide-exam
-e git+git://github.com/okfde/django-filingcabinet.git@master#egg=django-filingcabinet
-e git+git://github.com/okfde/django-amenities.git@master#egg=django-amenities
